Finish first part of scraper
Right now, the code is able to go to a main page containing many thumbnails
of books and then get the links to each of the more specific pages of those
books. On these more specific pages, I will grab the title and page count
and put that into a csv.