Account for dirty data situations
Not all books have a listed page count or author. I don't save any of the book's
data if there's no page count. If there's no author listed, I just write
N/A as the author.

Having an author isn't even pertinent to the project, I only need the page
count and book title but it was still a good learning experience.

Also, the page count comes accross with a ": " prepended so I use regular expressions
to extract only the page count (overkill but cool)